Move license issuing mutation definition to a hook
